Refactor Tickline Logic
Rather than going straight to a Line component within the XAxis, this introduces
a Tick hierarchy to parallel the Axis one (as Axis contains XAxis and YAxis, now
Tick will contain XTick and YTick).

This improves the semantics of the component signature, and gives a pleasant API
for working with it.

Additionally, something about the refactor fixes a judder that was noticed when
a chart with ticks rerendered.