XTicks rightside up again
During a merge-resolution, I upside-downed the XTick, so that it extended from
the top of the graph, instead of the bottom, this is now fixed, and the tests
are updated accordingly.

Additionally, to keep the HTML clean, all ticks are now wrapped in a <g> with
their corresponding label. Tests are updated to reflect this