Set compiled files as git binaries
Because dist/ and examples/ are checked into git, they appear in all diffs and
PRs. That's too bad, because they aren't meant for human consumption, nor would
a rational human be able to debug them.

This marks them as "binary" files in git, so that diffs will simply note that
they are changed, rather than providing an actual diff. This should markedly
reduce visual clutter when reviewing code.