Add theme editor to examples page
In order to see the impact changing the theme will have, this adds a form to
the examples page. This form sets the attributes for a theme and then that theme
is passed downward to all child examples, in order to show the impact it has.

Currently, this causes a serious performance hit during rapid changes to the
theme. That will need additional investigation. Additionally, the form has no
styling and is, thus, hideous.